+++ main.c
@@ -1,5 +1,4 @@
 #include "stdio.h"
-#include "source/config.h"
 #include "source/proxy.h"
 
 int main(int argc, char** argv){
